@@ -3,19 +3,18 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use JavaScript;
-use App\User;
+use Auth;
+use App\Models\User;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Auth;
 
 class UsersController extends Controller
 {
-       public function index()
+    public function index()
     {
         return view('users');
     }
 
-       public function registerUserView()
+    public function registerUserView()
     {
         return view('register');
     }
@@ -23,47 +22,47 @@ class UsersController extends Controller
        public function registerUser(Request $request)
     {
         $user = User::firstOrCreate(
-            ['email' => $request->input('password')], 
+            ['email' => $request->input('email')], 
             [
-                'username' => $request->input('userName'),
+                'username' => $request->input('username'),
                 'name' => $request->input('name'),
+				'role' => $request->input('role'),
+				'active' => $request->input('active'),
                 'password' => Hash::make($request->input('password')),
             ]
         );
         return $user;
     }
 
-       public function onlineUsers()
+    public function onlineUsers()
     {
         return User::where('id', '!=', Auth::id())->get();
     }
 
-       public function allUsers()
+    public function allUsers()
     {
         return User::all();
     }
 
-       public function getAuthUser()
+    public function getAuthUser()
     {
         return Auth::user();
     }
 
-       public function saveUser(Request $request , User $user)
+    public function saveUser(Request $request , User $user)
     {
         $user->name = $request->input('name');
+        $user->active = $request->input('active');
         $user->username = $request->input('username');
         $user->email = $request->input('email');
         $user->role = $request->input('role');
         $user->password = Hash::make($request->input('password'));
         $user->save();
-        return 'success';
+        return $user;
     }
 
-       public function getUser(User $user)
+    public function getUser(User $user)
     {
-        JavaScript::put([
-            'employee' => $user,
-        ]);
-        return view('user-single');
+        return $user;
     }
 }
