@@ -1,5 +1,7 @@
 <?php
 
+use Illuminate\Support\Facades\Broadcast;
+
 /*
 |--------------------------------------------------------------------------
 | Broadcast Channels
@@ -10,7 +12,6 @@
 | used to check if an authenticated user can listen to the channel.
 |
 */
-use Illuminate\Support\Facades\Auth;
 
 Broadcast::channel('App.User.{id}', function ($user, $id) {
     return (int) $user->id === (int) $id;
